Perubahan Tipe Data Nilai

+++ database/migrations/2021_09_19_192006_create_btps_table.php
@@ -26,7 +26,7 @@ class CreateBtpsTable extends Migration
             $table->string('nama');
             $table->enum('semester', ['1', '2']);
             $table->integer('kategori');
-            $table->integer('bobot');
+            $table->float('bobot');
             $table->timestamps();
         });
     }
