+++ database/migrations/2021_09_22_100017_create_bobotcpls_table.php
@@ -26,7 +26,7 @@ class CreateBobotcplsTable extends Migration
             $table->foreignIdFor(Cpmk::class);
             $table->foreignIdFor(Btp::class);
             $table->enum('semester', ['1', '2']);
-            $table->integer('bobot_cpl');
+            $table->float('bobot_cpl');
             $table->timestamps();
         });
     }
